requirements.txt
Flask==3.1.3
flask_login==0.6.3
flask_mail==0.10.0
pymysql==1.1.3
pypdf==6.7.1
python-dotenv==1.2.2
reportlab==4.4.10
Werkzeug==3.1.8
